+++ CHANGELOG.md
@@ -408,6 +408,7 @@
 * [X] 重构和添加`List.collapse`和`Any.expand`
 * [X] 重命名`String.splitMatched`和`String.splitToStrings`
 * [X] 使用java编写常量类。
+* [X] `Color`、`ColorService`、`String.toColor`、`String.toColorOrNull`以及注册默认的css颜色
 
 # 长期